// src/generated/types.gen.ts
export type Pullrequest = Omit<Object, 'type'> & {
// ...
    state?: 'OPEN' | 'MERGED' | 'DECLINED' | 'SUPERSEDED';
// ...
};